Create README.md
A Java-based graphical user interface application that mimics a food ordering system may be found in the FoodOrderingSystem repository. Customers can easily place their orders and choose goods from a menu by interacting with the user-friendly interface. The graphical layout provides a captivating user experience and was created with Java's Swing framework.

There are a lot of distinct foods in the application, and a checkbox represents each. The options open to users are "Rice & Beans," "Chapati," "Tea," "Chana Shamosa," "Thali," and "Ice cream." The software determines the total cost of the order when the desired goods have been chosen and the "ORDER" button is hit. A pop-up dialog box with the detailed receipt—along with the selected products and total amount—is then displayed to the user for their review.
This project is an excellent resource for beginners looking to learn programming in Java. Its clean and well-organized codebase serves as a foundation for understanding event handling, component layout, and user input processing. Developers can use this project as a starting point to expand and enhance the features of the food ordering system, potentially incorporating additional functionality such as payment processing, order history management, or user authentication.